initial Commit for Nashville housing
In this project we cleaned and explored a Nashville housing dataset from Kaggle using SQL. The dataset documents the housing market in Nashville in particular the value and selling price of a house. The purpose of this project was to clean our data in order to perform our data exploration and analysis. Enjoy!